Add job statistics gathering, yay!
No test for it yet, though. That's for later. It's friday, after all!

+++ db/migrate/20191204224546_create_job_executions.rb
@@ -2,7 +2,7 @@ class CreateJobExecutions < ActiveRecord::Migration[6.0]
   def change
     create_table :job_executions do |t|
       t.string :provider_job_id, null: false
-      t.integer :duration, null: false
+      t.numeric :duration, null: false
 
       t.timestamps
     end
